+++ build.gradle
@@ -27,38 +27,39 @@ sourceSets {
         }
         resources {
             srcDirs = ["src/main/resources"]
-        }
-    }
- }
+		}
+	}
+}
 
 dependencies {
     def mapsforgeVersion = "+"
     // https://mvnrepository.com/artifact/org.eclipse.jetty.http2/http2-server
-    def jettyVersion = "11.0.21"
+    def jettyVersion = "11.0.23"
     def slf4j = "+"
 
     implementation "org.mapsforge:mapsforge-core:${mapsforgeVersion}"
     implementation "org.mapsforge:mapsforge-map:${mapsforgeVersion}"
-    implementation "org.mapsforge:mapsforge-map-reader:${mapsforgeVersion}"
-    runtimeOnly "org.mapsforge:mapsforge-themes:${mapsforgeVersion}"
     implementation "org.mapsforge:mapsforge-map-awt:${mapsforgeVersion}"
+    implementation "org.mapsforge:mapsforge-map-reader:${mapsforgeVersion}"
+    runtimeOnly    "org.mapsforge:mapsforge-themes:${mapsforgeVersion}"
 
     implementation "org.eclipse.jetty:jetty-server:${jettyVersion}"
-    runtimeOnly "org.eclipse.jetty.http2:http2-server:${jettyVersion}"
+    runtimeOnly    "org.eclipse.jetty.http2:http2-server:${jettyVersion}"
     implementation "org.eclipse.jetty:jetty-util:${jettyVersion}"
-    implementation "org.eclipse.jetty:jetty-xml:${jettyVersion}"
+	implementation "org.eclipse.jetty:jetty-xml:${jettyVersion}"
     implementation "org.eclipse.jetty.toolchain:jetty-jakarta-servlet-api:+"
 
-    runtimeOnly "org.slf4j:slf4j-reload4j:${slf4j}"
+    runtimeOnly    "org.slf4j:slf4j-reload4j:${slf4j}"
     implementation "org.slf4j:jul-to-slf4j:${slf4j}"
-    implementation "org.slf4j:slf4j-api:${slf4j}"
-    
+	implementation "org.slf4j:slf4j-api:${slf4j}"
+
     implementation "commons-cli:commons-cli:+"
     implementation "org.apache.commons:commons-lang3:+"
     implementation "com.github.marschall:memoryfilesystem:+"
-    
-    runtimeOnly "org.xerial:sqlite-jdbc:+"
-    runtimeOnly "net.sf.kxml:kxml2:+"   
+//  implementation "com.google.jimfs:jimfs:+"  <-- Alternative memory filesystem
+
+    runtimeOnly    "org.xerial:sqlite-jdbc:+"
+    runtimeOnly    "net.sf.kxml:kxml2:+" 
 }
 
 application {
