+++ config/server.properties
@@ -3,14 +3,7 @@
 # value might be quoted
 # file path must use slash (and not backslash) even on Windows system
 
-renderer=database
-# https://ftp.gwdg.de/pub/misc/openstreetmap/openandromaps/mapsV5/
-mapfiles=D:/mapsforgesrv/maps
-language=en
-# https://drive.google.com/drive/folders/0BxphPoRgwhnoWkRoTFhMbTM3RDA?resourcekey=0-wRe5bWl96pwvQ9tAfI9cQg
-demfolder=D:/mapsforgesrv/dem/dtm-srtm1
-outofrange_tms=https://a.tile.openstreetmap.fr/osmfr/{z}/{x}/{y}.png
-worldmap=true
 cache-control=2592000
+outofrange_tms=https://a.tile.openstreetmap.fr/osmfr/{z}/{x}/{y}.png
 # https://eclipse.dev/jetty/javadoc/jetty-11/org/eclipse/jetty/server/CustomRequestLog.html
 requestlog-format=%{yyyy-MM-dd_HH:mm:ss.SSS ZZZ|CET}t %{client}a %{ms}T '%r' %s %O '%{User-Agent}i'
