--- config/styles/default.properties
@@ -1,18 +1,0 @@
-# server configuration file to be at root of 'config directory' specified with arg -c
-# 'config directory' maybe be absolute or relative to user.dir
-# value might be quoted
-# file path must use slash (and not backslash) even on Windows system
-
-# diffuselight(angle)	  https://javadoc.io/static/org.mapsforge/mapsforge-map/0.21.0/org/mapsforge/map/layer/hills/DiffuseLightShadingAlgorithm.html
-# simple(linearity,scale) https://javadoc.io/static/org.mapsforge/mapsforge-map/0.21.0/org/mapsforge/map/layer/hills/SimpleShadingAlgorithm.html
-hillshading-algorithm=simple(0.4,1)
-# Increase (>1) or decrease (<1) the hillshading magnitude relative to the value set in themes [0-4]
-# default: 64 [0-255] - 64 in theme
-hillshading-magnitude=2
-contrast-stretch=0
-gamma-correction=1
-device-scale=1.2
-user-scale=1
-text-scale=1
-symbol-scale=1
-line-scale=1
