+++ config/tasks/active.properties
@@ -1,0 +1,44 @@
+# server configuration file to be at root of 'config directory' specified with arg -c
+# 'config directory' maybe be absolute or relative to user.dir
+# value might be quoted
+# file path must use slash (and not backslash) even on Windows system
+
+renderer=database
+# https://ftp.gwdg.de/pub/misc/openstreetmap/openandromaps/mapsV5/
+mapfiles=D:/mapsforgesrv/maps
+language=en
+# https://drive.google.com/drive/folders/0BxphPoRgwhnoWkRoTFhMbTM3RDA?resourcekey=0-wRe5bWl96pwvQ9tAfI9cQg
+demfolder=D:/mapsforgesrv/dem/dtm-srtm1
+worldmap=true
+
+# https://github.com/FrankSchoeneck/active_RT5
+themefile=D:/mapsforgesrv/themes/active/active RT5.xml
+style=Style_Summer
+
+# L_hikingtrails		Waymarked Trails
+# L_trail_markings		Trail markings
+# L_hikingnodes			Hiking Nodes
+# L_contourlines		Contour lines
+# L_boundaries_areas		Boundaries + protected areas
+# L_landscapefeat		landscape features
+# L_publictransport		Public transport
+# L_parking			Parking
+# L_restaurant			Restaurants + Cafés
+# L_hotel			Hotels + Accommodation
+# L_emergency_access_point	Emergency access points
+
+overlays=L_hikingtrails,L_trail_markings,L_hikingnodes,L_contourlines,L_boundaries_areas,L_landscapefeat,L_publictransport,L_parking,L_restaurant,L_hotel,L_emergency_access_point
+
+# diffuselight(angle)	  https://javadoc.io/static/org.mapsforge/mapsforge-map/0.21.0/org/mapsforge/map/layer/hills/DiffuseLightShadingAlgorithm.html
+# simple(linearity,scale) https://javadoc.io/static/org.mapsforge/mapsforge-map/0.21.0/org/mapsforge/map/layer/hills/SimpleShadingAlgorithm.html
+hillshading-algorithm=simple(0.4,1)
+# Increase (>1) or decrease (<1) the hillshading magnitude relative to the value set in themes [0-4]
+# default: 64 [0-255] - 64 in theme
+hillshading-magnitude=2
+contrast-stretch=0
+gamma-correction=1
+device-scale=1
+user-scale=0.8
+text-scale=1.2
+symbol-scale=1
+line-scale=1
