+++ config/tasks/hillshading.properties
@@ -1,0 +1,8 @@
+# server configuration file to be at root of 'config directory' specified with arg -c
+# 'config directory' maybe be absolute or relative to user.dir
+# value might be quoted
+# file path must use slash (and not backslash) even on Windows system
+
+demfolder=D:/Landkarten/DEM/Hgt3
+hillshading-algorithm=simple(0.4,1)
+hillshading-magnitude=2
