+++ config/tasks/nono.properties
@@ -1,0 +1,56 @@
+# server configuration file to be at root of 'config directory' specified with arg -c
+# 'config directory' maybe be absolute or relative to user.dir
+# value might be quoted
+# file path must use slash (and not backslash) even on Windows system
+
+renderer=database
+# https://ftp.gwdg.de/pub/misc/openstreetmap/openandromaps/mapsV5/
+mapfiles=D:/mapsforgesrv/maps
+language=en
+# https://drive.google.com/drive/folders/0BxphPoRgwhnoWkRoTFhMbTM3RDA?resourcekey=0-wRe5bWl96pwvQ9tAfI9cQg
+demfolder=D:/mapsforgesrv/dem/dtm-srtm1
+worldmap=true
+
+# https://github.com/nono303/mapsforge-theme-elevate
+themefile=D:/mapsforgesrv/themes/nono/Elevate-NoNo.xml
+style=elv-hiking
+
+# elv-h_routes		[R] Hiking routes
+# elv-h_s_routes	[R] Emphasized routes
+# elv-waymarks		[R] Waymarks
+# elv-road_surfaces-h	[W] Extended road surfaces
+# elv-acc_allowed-h	[W] Marking of permitted footways
+# elv-settlements-h	[P] Settlement names
+# elv-landscapefeat-h	[P] Landscape features
+# elv-amenities-h	[P] Amenities
+# elv-barriers-h	[P] Barriers
+# elv-sports-h		[P] Sports
+# elv-emergency-h	[P] Emergency
+# elv-accommodation-h	[P] Accommodation
+# elv-restaurants-h	[P] Restaurants + bars
+# elv-shops-h		[P] Shops + services
+# elv-tourism-h		[P] Tourism + culture
+# elv-pubtrans-h	[P] Public transport
+# elv-car-h		[P] Car
+# elv-buildings-h	[P] Special buildings
+# elv-borders-h		[A] Borders + special areas
+# elv-background-h	[A] Map background
+# elv-pt_network false	[R] Public transport network
+# elv-smallways		[W] Paths + tracks
+
+# disabled: elv-pt_network,elv-h_s_routes,elv-smallways
+overlays=elv-h_routes,elv-waymarks,elv-road_surfaces-h,elv-acc_allowed-h,elv-settlements-h,elv-landscapefeat-h,elv-amenities-h,elv-barriers-h,elv-sports-h,elv-emergency-h,elv-accommodation-h,elv-restaurants-h,elv-shops-h,elv-tourism-h,elv-pubtrans-h,elv-car-h,elv-buildings-h,elv-borders-h,elv-background-h
+
+# diffuselight(angle)	  https://javadoc.io/static/org.mapsforge/mapsforge-map/0.21.0/org/mapsforge/map/layer/hills/DiffuseLightShadingAlgorithm.html
+# simple(linearity,scale) https://javadoc.io/static/org.mapsforge/mapsforge-map/0.21.0/org/mapsforge/map/layer/hills/SimpleShadingAlgorithm.html
+hillshading-algorithm=simple(0.4,1)
+# Increase (>1) or decrease (<1) the hillshading magnitude relative to the value set in themes [0-4]
+# default: 64 [0-255] - 64 in theme
+hillshading-magnitude=2
+contrast-stretch=0
+gamma-correction=1
+device-scale=1.2
+user-scale=1
+text-scale=1
+symbol-scale=1
+line-scale=1
