+++ config/tasks/tapiola.properties
@@ -1,0 +1,80 @@
+# server configuration file to be at root of 'config directory' specified with arg -c
+# 'config directory' maybe be absolute or relative to user.dir
+# value might be quoted
+# file path must use slash (and not backslash) even on Windows system
+
+renderer=database
+# https://ftp.gwdg.de/pub/misc/openstreetmap/openandromaps/mapsV5/
+mapfiles=D:/mapsforgesrv/maps
+language=en
+# https://drive.google.com/drive/folders/0BxphPoRgwhnoWkRoTFhMbTM3RDA?resourcekey=0-wRe5bWl96pwvQ9tAfI9cQg
+demfolder=D:/mapsforgesrv/dem/dtm-srtm1
+worldmap=true
+
+# https://github.com/TapioOfTapiola/oam_theme_tapiola/releases
+themefile=D:/mapsforgesrv/themes/tapiola/Tapiola.xml
+style=allfeats
+
+# L_hiking_trails 	Hiking network
+# L_hiking_nodes 	Hiking Nodes
+# L_cycling 		Cycling network
+# L_mtb_trails 		MTB trails
+# L_winter_sports 	Winter sports
+# L_railway 		Railways
+# L_highway 		Road network
+# L_road_captions 	Road names
+# L_grades 		Track grade codes
+# L_surface_text 	Surface as text
+# L_unpaved 		Unpaved road indicators
+# L_access 		Access restriction indicators
+# L_buildings 		Buildings
+# L_building_type 	Colored schools, hospitals...
+# L_house_numbers 	House numbers
+# L_house_nr_early 	House numbers earlier
+# L_cities_towns 	City/Town/Village names
+# L_cities_towns_sub 	Suburb/Industrial area names
+# L_publictransport 	Public transport, traffic
+# L_barriers 		Barriers and gates
+# L_parking 		Parking
+# L_parking_lowzoom 	Display parking earlier
+# L_sports 		Sports
+# L_restaurants 	Restaurants and Cafés
+# L_shops 		Shops and services
+# L_accommodation 	Hotels and Accommodation
+# L_pharmacies 		Pharmacies and hospitals
+# L_emergency 		Emergency (police, firemen, ...)
+# L_architecture 	Architecture, religious, museums
+# L_attractions 	Tourist/Attractions
+# L_playground 		Playgrounds
+# L_toilets 		Toilets
+# L_benches		Benches and tables
+# L_refuge 		Refuge
+# L_hunting_stand 	Hunting stands
+# L_misc 		Miscellaneous
+# L_water 		Spring, water
+# L_landscapefeat 	Peaks, landscape features
+# L_forest 		Forest, grass, crops
+# L_contourlines 	Contour lines
+# L_protected_area 	Protected areas
+# L_mil_area 		Military area
+# L_boundaries_areas 	Administrative boundaries
+# L_mast_communication	Cellular tower
+# L_electricity 	Electricity
+# L_white_overlay 	Map gamma
+
+# disabled: L_cycling,L_mtb_trails,L_white_overlay,L_protected_area
+overlays=L_hiking_trails,L_hiking_nodes,L_winter_sports,L_railway,L_highway,L_road_captions,L_grades,L_surface_text,L_unpaved,L_access,L_buildings,L_building_type,L_house_numbers,L_house_nr_early,L_cities_towns,L_cities_towns_sub,L_publictransport,L_barriers,L_parking,L_parking_lowzoom,L_sports,L_restaurants,L_shops,L_accommodation,L_pharmacies,L_emergency,L_architecture,L_attractions,L_playground,L_toilets,L_benches,L_refuge,L_hunting_stand,L_misc,L_water,L_landscapefeat,L_forest,L_contourlines,L_mil_area,L_boundaries_areas,L_mast_communication,L_electricity
+
+# diffuselight(angle)	  https://javadoc.io/static/org.mapsforge/mapsforge-map/0.21.0/org/mapsforge/map/layer/hills/DiffuseLightShadingAlgorithm.html
+# simple(linearity,scale) https://javadoc.io/static/org.mapsforge/mapsforge-map/0.21.0/org/mapsforge/map/layer/hills/SimpleShadingAlgorithm.html
+hillshading-algorithm=simple(0.4,1)
+# Increase (>1) or decrease (<1) the hillshading magnitude relative to the value set in themes [0-4]
+# default: 64 [0-255] - 64 in theme
+hillshading-magnitude=2
+contrast-stretch=0
+gamma-correction=1
+device-scale=1.2
+user-scale=1
+text-scale=1
+symbol-scale=1
+line-scale=1
