+++ settings.gradle
@@ -13,4 +13,4 @@ plugins {
     id 'org.gradle.toolchains.foojay-resolver-convention' version '0.8.0'
 }
 
-rootProject.name = 'mapsforgesrv'
+rootProject.name = 'mapsforgesrv_styles'
