+++ src/main/java/com/telemaxx/mapsforgesrv/MapsforgeConfig.java
@@ -2,12 +2,8 @@ package com.telemaxx.mapsforgesrv;
 
 import java.io.File;
 import java.io.FilenameFilter;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
@@ -23,20 +19,14 @@ public class MapsforgeConfig extends PropertiesParser{
 
 	private CommandLine configCmd;
 
-	private ArrayList<File> mapFiles = null;
-	private File demFolder = null;
 	private long cacheControl;
-	private String rendererName = null;
 	private String outOfRangeTms = null;
-	private boolean appendWorldMap;
 	private boolean acceptTerminate;
-	private Map<String, MapsforgeStyleConfig> stylesConfig;
+	private Map<String, MapsforgeTaskConfig> tasksConfig;
 	private String configDirectory = null;
-	private String preferredLanguage = null;
 	private String requestLogFormat = null;
 
-	private final static String styleNameRegex = "^[0-9a-z._-]+$"; //$NON-NLS-1$
-	private final static Pattern styleNameRegexPattern = Pattern.compile(styleNameRegex);
+	private final static String taskFileNameRegex = "^[0-9a-z._-]+.properties$"; //$NON-NLS-1$
 
 	private final static Logger logger = LoggerFactory.getLogger(MapsforgeConfig.class);
 
@@ -52,8 +42,12 @@ public class MapsforgeConfig extends PropertiesParser{
 		Options options = new Options();
 		options.addOption(Option.builder("c") //$NON-NLS-1$
 				.longOpt("config") //$NON-NLS-1$
-				.desc("Config directory including at least "+FILECONFIG_SERVER+", "+FILECONFIG_JETTY+", "+DIRCONFIG_STYLE+FILECONFIG_DEFAULTSTYLE) //$NON-NLS-1$
-				.required(true).hasArg(true).build());
+				.desc("Config directory including at least "+FILECONFIG_SERVER+", "+FILECONFIG_JETTY+", "+FILECONFIG_JETTY_THREADPOOL+", "+DIRCONFIG_TASK+FILECONFIG_DEFAULTTASK) //$NON-NLS-1$
+				.required(false).hasArg(true).build());
+		options.addOption(Option.builder("h") //$NON-NLS-1$
+				.longOpt("help") //$NON-NLS-1$
+				.desc("Print this help text and exit") //$NON-NLS-1$
+				.required(false).hasArg(false).build());
 		CommandLineParser parser = new DefaultParser();
 		HelpFormatter formatter = new HelpFormatter();
 		formatter.setWidth(132);
@@ -72,21 +66,15 @@ public class MapsforgeConfig extends PropertiesParser{
 		if (config != null) {
 			config = config.trim();
 			if (new File(config).isDirectory()) {
-				configDirectory = config+"/";
-				if (new File(configDirectory+FILECONFIG_SERVER).isFile()) {
-					readConfig(configDirectory+FILECONFIG_SERVER);
-				} else {
-					logger.error("Server config file '"+configDirectory+FILECONFIG_SERVER+"' doesn't exist: exiting"); //$NON-NLS-1$
-					System.exit(1);
+				configDirectory = config+System.getProperty("file.separator");
+				String[] configFiles = {FILECONFIG_SERVER, FILECONFIG_JETTY, FILECONFIG_JETTY_THREADPOOL, DIRCONFIG_TASK+FILECONFIG_DEFAULTTASK};  
+				for (String configFile : configFiles) {
+					if (!new File(configDirectory+configFile).isFile()) {
+						logger.error("Default task config file '"+configDirectory+configFile+"' doesn't exist: exiting"); //$NON-NLS-1$
+						System.exit(1);
+					}
 				}
-				if (!new File(configDirectory+FILECONFIG_JETTY).isFile()) {
-					logger.error("Jetty config file '"+configDirectory+FILECONFIG_JETTY+"' doesn't exist: exiting"); //$NON-NLS-1$
-					System.exit(1);
-				}
-				if (!new File(configDirectory+DIRCONFIG_STYLE+FILECONFIG_DEFAULTSTYLE).isFile()) {
-					logger.error("Default style config file '"+configDirectory+DIRCONFIG_STYLE+FILECONFIG_DEFAULTSTYLE+"' doesn't exist: exiting"); //$NON-NLS-1$
-					System.exit(1);
-				}
+				readConfig(configDirectory+FILECONFIG_SERVER);
 			} else {
 				logger.error("Config directory '"+config+"' set with -c is not a directory: exiting"); //$NON-NLS-1$
 				System.exit(1);
@@ -103,91 +91,37 @@ public class MapsforgeConfig extends PropertiesParser{
 
 	private void initConfig() throws Exception {
 		logger.info("################## SERVER CONFIG ##################");
-		parseMapFiles();
-		preferredLanguage = parseString(null, "language", null, "Preferred map language"); //$NON-NLS-1$ //$NON-NLS-2$
-		demFolder = parseFile("demfolder", FOLDER, true, "DEM", "undefined");
-		rendererName = parseString(DEFAULT_RENDERER, "renderer", AUTHORIZED_RENDERER, "Renderer algorithm"); //$NON-NLS-1$ //$NON-NLS-2$
 		cacheControl = (long) parseNumber(DEFAULT_CACHECONTROL, "cache-control", 0, null, "Browser cache ttl",false); //$NON-NLS-1$ //$NON-NLS-2$
 		outOfRangeTms = parseString(null, "outofrange_tms", null, "Out of range TMS url"); //$NON-NLS-1$ //$NON-NLS-2$
-		appendWorldMap = parseHasOption("worldmap", "Append built-in world map");
 		acceptTerminate = parseHasOption("terminate", "Accept terminate request");
 		requestLogFormat = parseString("%{client}a - %u %t '%r' %s %O '%{Referer}i' '%{User-Agent}i' '%C'", "requestlog-format", null, "Request log format"); //$NON-NLS-1$ //$NON-NLS-2$
-		parseStyles();
+		parseTasks();
 	}
 	
 	/*
 	 * PARSERS
 	 */
 	
-	private void checkStyleName(String styleName) {
-		if(!styleNameRegexPattern.matcher(styleName).find()) {
-			logger.error("properties files for style '"+styleName+"' had not a correct name '"+styleNameRegex+"'"); //$NON-NLS-1$
-			System.exit(-1);
-		}
-	}
-	
-	private void parseStyles() throws Exception {
-		File[] styleFiles = new File(configDirectory+"/styles").listFiles();
-		if(styleFiles.length == 0) {
-			logger.error(configDirectory+"/styles doesn't contain any properties files"); //$NON-NLS-1$
+	private void parseTasks() throws Exception {
+		FilenameFilter filenameFilter = new FilenameFilter() {
+			    public boolean accept(File dir, String name) {
+			    	return name.matches(taskFileNameRegex);
+			    }};
+		File[] taskFiles = new File(configDirectory+"/tasks").listFiles(filenameFilter);
+		if(taskFiles.length == 0) {
+			logger.error(configDirectory+"/tasks doesn't contain any properties files named "+taskFileNameRegex); //$NON-NLS-1$
 			System.exit(-1);
 		} else {
-			stylesConfig = new HashMap<String, MapsforgeStyleConfig>();
-			MapsforgeStyleConfig msc;
-			for (File styleFile : styleFiles) { 
-				String styleName = styleFile.getName().replaceFirst("[.][^.]+$", ""); //$NON-NLS-1$
-				checkStyleName(styleName);
-				msc = new MapsforgeStyleConfig(styleName, styleFile);
-				stylesConfig.put(styleName, msc);
+			tasksConfig = new HashMap<String, MapsforgeTaskConfig>();
+			MapsforgeTaskConfig msc;
+			for (File taskFile : taskFiles) { 
+				String taskFileName = taskFile.getName();
+				logger.info("taskFileName="+taskFileName);
+				String taskName = taskFileName.replaceFirst("[.][^.]+$", ""); //$NON-NLS-1$
+				logger.info("taskName="+taskName);
+				msc = new MapsforgeTaskConfig(taskName, taskFile);
+				tasksConfig.put(taskName, msc);
 			}
-		}
-	}
-	
-	private void parseMapFiles() throws Exception {
-		String msgHeader = parsePadMsg("Map file(s)"); //$NON-NLS-1$
-		String mapFilePathsString = retrieveConfigValue("mapfiles"); //$NON-NLS-1$
-		if (mapFilePathsString != null) {
-			String[] mapFilePaths = mapFilePathsString.trim().split(","); //$NON-NLS-1$ //$NON-NLS-2$
-			mapFiles = new ArrayList<File>();
-			List<File> mapsErr = new ArrayList<File>();
-			for (String path : mapFilePaths) {
-				File file = new File(path.trim());
-				if (file.exists()) {
-					if (file.isFile()) {
-						mapFiles.add(file);
-					} else if (file.isDirectory()) {
-						for (File mapfile : file.listFiles(new FilenameFilter() {
-							@Override
-							public boolean accept(File dir, String name) {
-								return name.endsWith(".map");
-							}
-						})) {
-							mapFiles.add(mapfile);
-						}
-					}
-				}
-			}
-			mapFiles.forEach(mapFile -> {
-				if (!mapFile.isFile())
-					mapsErr.add(mapFile);
-			});
-			String mapFilesString;
-			if (mapsErr.size() > 0) {
-				mapFiles.removeAll(mapsErr);
-				mapFilesString = mapFiles.stream().map(File::getPath).collect(Collectors.joining(","));
-				String cnxNotAuth = "{" + mapsErr.stream().map(File::getPath).collect(Collectors.joining(",")) + "} not existing"; //$NON-NLS-2$ //$NON-NLS-3$
-				if (mapFilePaths.length == 0) {
-					parseError(msgHeader, cnxNotAuth, "{" + mapFilesString + "}");
-				} else {
-					logger.info(msgHeader + ": defined [{" + mapFilesString + "}] - warn " + cnxNotAuth); //$NON-NLS-1$
-				}
-			} else {
-				mapFilesString = mapFiles.stream().map(File::getPath).collect(Collectors.joining(","));
-				logger.info(msgHeader + ": defined [{" + mapFilesString + "}]"); //$NON-NLS-1$
-			}
-		} else {
-			logger.error(msgHeader + ": exiting - no file(s) specified"); //$NON-NLS-1$
-			System.exit(1);
 		}
 	}
 
@@ -198,19 +132,7 @@ public class MapsforgeConfig extends PropertiesParser{
 	public String getConfigDirectory() {
 		return configDirectory;
 	}
-	
-	public File getDemFolder() {
-		return this.demFolder;
-	}
 
-	public List<File> getMapFiles() {
-		return this.mapFiles;
-	}
-	
-	public boolean getAppendWorldMap() {
-		return this.appendWorldMap;
-	}
-	
 	public boolean getAcceptTerminate() {
 		return this.acceptTerminate;
 	}
@@ -218,38 +140,29 @@ public class MapsforgeConfig extends PropertiesParser{
 	public long getCacheControl() {
 		return this.cacheControl;
 	}
-	
+
 	public String getOutOfRangeTms() {
 		return this.outOfRangeTms;
 	}
-	
-	public String getRendererName() {
-		return this.rendererName;
-	}
-	
-	public String getPreferredLanguage() {
-		return this.preferredLanguage;
-	}
-	
+
 	public String getRequestLogFormat() {
 		return requestLogFormat;
 	}
 
-	
-	public Map<String, MapsforgeStyleConfig> getStylesConfig() {
-		return stylesConfig;
+	public Map<String, MapsforgeTaskConfig> getTasksConfig() {
+		return tasksConfig;
 	}
-	
-	public MapsforgeStyleConfig getStylesConfig(String style) throws Exception {
+
+	public MapsforgeTaskConfig getTaskConfig(String task) throws Exception {
 		try {
-			return stylesConfig.get(style);
+			return tasksConfig.get(task);
 		} catch(Exception e) {
-			throw new Exception("Style '"+style+"' don't exist");
+			throw new Exception("Task '"+task+"' don't exist");
 		}
 	}
 
-	public MapsforgeStyleConfig getDefaultStyle() throws Exception {
-		return getStylesConfig("default");
+	public MapsforgeTaskConfig getDefaultConfig() throws Exception {
+		return getTaskConfig("default");
 	}
 
 }
